Fix: 자료형 오류 수정

+++ server/services/assignment_service.py
@@ -134,7 +134,7 @@ def edit_assignment(as_no,limit_time, as_name, as_type, keyword, prob_translang_
             #json region을 dict로 변환
             region = region.replace("'",'"')
             region = json.loads(region)
-            split_url=split_wav_save2(prob_sound_path,int(region["start"]),int(region["end"]))
+            split_url=split_wav_save2(prob_sound_path,float(region["start"]),float(region["end"]))
             mapping_sst_user(assignment_to_edit.assignment_no, split_url,user_info)
             task = do_stt_work.delay(filename=split_url,locale=prob_translang_source)
             pr = Prob_region(assignment_no=assignment_to_edit.assignment_no,region_index=region["index"],start=region["start"],end=region["end"],upload_url=split_url, job_id=task.id)
